updated tensor colums to use decimals instead of integer

+++ backend/db/migrate/20200329015541_add_tensor_flow_columns.rb
@@ -1,7 +1,7 @@
 class AddTensorFlowColumns < ActiveRecord::Migration[5.2]
   change_table :clothings do | t |
-    t.integer :left_offset_percent
-    t.integer :right_offset_percent
-    t.integer :top_offset_percent
+    t.decimal :left_offset_percent
+    t.decimal :right_offset_percent
+    t.decimal :top_offset_percent
   end
 end
